Remove Hex.API.Utils
